Dedupe `make_lsp_adapter_delegate`
